# artifacts/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Talkam Liberia API
  version: "1.0.0"
  description: |
    REST API for Talkam Liberia mobile/web apps. Supports JWT + anonymous ephemeral tokens.
servers:
  - url: https://api.talkamliberia.org/v1
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      summary: Register a user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { description: Validation error }
  /auth/login:
    post:
      summary: Login and obtain JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Auth token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401': { description: Invalid credentials }
  /auth/anonymous-start:
    post:
      summary: Issue ephemeral anonymous token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnonymousStartRequest'
      responses:
        '200':
          description: Ephemeral token details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnonymousStartResponse'
...
  /ngos/register:
    post:
      summary: Register NGO partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NGORegisterRequest'
      responses:
        '201': { description: NGO pending approval }
  /ngos/{id}/dashboard:
    get:
      summary: NGO dashboard data
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Dashboard payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NGODashboardResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [full_name, phone, password]
      properties:
        full_name: { type: string }
        phone: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password }
        language: { type: string, default: en-LR }
    LoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone: { type: string }
        password: { type: string, format: password }
        device_id: { type: string }
    AnonymousStartRequest:
      type: object
      properties:
        device_hash: { type: string }
        county: { type: string }
        capabilities: { type: array, items: { type: string } }
    AnonymousStartResponse:
      type: object
      properties:
        token: { type: string }
        expires_in: { type: integer }
        offline_queue_limit: { type: integer }
    AuthResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer }
        roles: { type: array, items: { type: string } }
...
    ReportSummary:
      type: object
      properties:
        id: { type: string }
        summary: { type: string }
        category: { type: string }
        county: { type: string }
        severity: { type: string }
        status: { type: string }
...
    TimelineEvent:
      type: object
      properties:
        at: { type: string, format: date-time }
        action: { type: string }
        actor: { type: string }
        notes: { type: string }
...
    NGORegisterRequest:
      type: object
      properties:
        name: { type: string }
        contact_email: { type: string }
        phone: { type: string }
        mission_focus: { type: array, items: { type: string } }
        counties: { type: array, items: { type: string } }
    NGODashboardResponse:
      type: object
      properties:
        pending_reports: { type: array, items: { $ref: '#/components/schemas/ReportSummary' } }
        follow_ups: { type: array, items: { $ref: '#/components/schemas/TimelineEvent' } }
        exports:
          type: object
          properties:
            csv_url: { type: string }
            pdf_url: { type: string }
...